Improved einsum performance by deferring inner-loop to sum_of_products function

+++ benches/perfprofiler/util.py
@@ -34,7 +34,7 @@ def get_class_from_method(method) -> str:
     return method.__qualname__.split(".")[0]
 
 
-def rand_ndarrays_with_shape(shapes: list[tuple[int, ...]], dtype: str = "float32") -> list[np.ndarray]:
+def rand_ndarrays_with_shape(shapes: list[tuple[int, ...]], dtype: str = "int32") -> list[np.ndarray]:
     return [np.random.rand(*shape).astype(dtype) for shape in shapes]
